requirements.txt
Flask==2.3.2
PyCryptodome==3.18.0
mysql-connector-python==8.2.0
bcrypt==4.1.2
reportlab==4.0.8
qrcode==7.4.2
